requirements.txt
google-adk
pinecone
requests
pandas
fastapi
uvicorn
rich
python-dotenv
python-multipart